requirements.txt
discord.py
scikit-learn
dotenv